Extracted loadPointCloud function to util.
